Add parser.py
